fix: Extract graphql-jpa-query-web module from starter
