+++ graphql-jpa-query-boot-starter/pom.xml
@@ -24,6 +24,11 @@
     </dependency>
 
     <dependency>
+      <groupId>com.introproventures</groupId>
+      <artifactId>graphql-jpa-query-web</artifactId>
+    </dependency>
+
+    <dependency>
       <groupId>org.springframework.boot</groupId>
       <artifactId>spring-boot-starter</artifactId>
     </dependency>
