+++ graphql-jpa-query-boot-starter/src/test/java/com/introproventures/graphql/jpa/query/starter/model/Author.java
@@ -14,27 +14,24 @@
  * limitations under the License.
  */
 
-package com.introproventures.graphql.jpa.query.web.model;
+package com.introproventures.graphql.jpa.query.starter.model;
+
+import java.util.Collection;
 
 import javax.persistence.Entity;
-import javax.persistence.EnumType;
-import javax.persistence.Enumerated;
 import javax.persistence.Id;
-import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
 
 import lombok.Data;
 
 @Data
 @Entity
-public class Book {
+public class Author {
 	@Id
 	Long id;
 
-	String title;
+	String name;
 
-	@ManyToOne
-	Author author;
-
-	@Enumerated(EnumType.STRING)
-	Genre genre;
+	@OneToMany(mappedBy="author")
+	Collection<Book> books;
 }
