+++ graphql-jpa-query-dependencies/pom.xml
@@ -59,6 +59,11 @@
                 <artifactId>graphql-jpa-query-autoconfigure</artifactId>
                 <version>${project.version}</version>
             </dependency>
+            <dependency>
+                <groupId>com.introproventures</groupId>
+                <artifactId>graphql-jpa-query-web</artifactId>
+                <version>${project.version}</version>
+            </dependency>
         </dependencies>
     </dependencyManagement>
