+++ graphql-jpa-query-web/pom.xml
@@ -1,0 +1,24 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+  <parent>
+    <groupId>com.introproventures</groupId>
+    <artifactId>graphql-jpa-query-build</artifactId>
+    <version>0.3.13-SNAPSHOT</version>
+    <relativePath>../graphql-jpa-query-build</relativePath>
+  </parent>
+  <artifactId>graphql-jpa-query-web</artifactId>
+
+  <dependencies>
+    <dependency>
+      <groupId>com.introproventures</groupId>
+      <artifactId>graphql-jpa-query-schema</artifactId>
+      <optional>true</optional>
+    </dependency>  
+    <dependency>
+      <groupId>org.springframework.boot</groupId>
+      <artifactId>spring-boot-starter-web</artifactId>
+      <optional>true</optional>
+    </dependency>  
+  </dependencies>  
+  
+</project>
