+++ graphql-jpa-query-web/src/main/resources/META-INF/spring.factories
@@ -1,0 +1,2 @@
+org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
+	com.introproventures.graphql.jpa.query.web.GraphQLControllerAutoConfiguration
